Move accordions description to MaxWidthTextBlock component
